{dutch}.json
I've tried to translate the English words into Dutch so specific and precise as possible for not every Dutchman understands English as well as most do. Hopefully the translation is better to comprehend for them.

Some words are indecisevly difficult to translate. This is because some words are best comprehended when kept in English. For the most part this is because sometimes these words are to technical.
But than again also because they are well understood by most Dutch people and is part of the anglicization of our language. 

Some of the words giving some difficulty may be:

summary: overzicht/ samengevat?/ sammenvatting?/ beknopt?
you: u/ je? (using u for being polite)
Wallet: portemonnee/ portefeuille?
Merchants: handelaars / kooplieden?/ 
update: bijgewerkt/ upgedate?
input: input/ ingang?
output: output/ uitgang?
Fee: fee/ vergoeding?/ beloning?
Sweep: Schoonvegen/ wegvegen?/ vegen?
Miners: Miners /mijnwerkers?
review: goedkeuren/ beoordelen
Unspend: Onbesteed(baar)/ niet-uitgegeven?/ onuitgegeven?

The first words given are the ones being used in this translation.

For me being a new Blockchain wallet user, I'll still getting used to the website and maybe stumble upon strange Dutch sentences. I will watch it and then review and change the dutch.json language file.

It's not really clear to me what the precise conventions are, when capital letters are used and with what words only.

However this translation may not be complete and may have to be revised once again.

Hopefully now the text is for every Dutchman to comprehend from young to old, new or experienced. 

Ofcourse I'm open to suggestions and any questions at all.